src_django/requirements.txt
django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
redis==5.0.1
debugpy
